server/requirements.txt
django
djangorestframework
django-celery
geopy